[milestone] It compiles!
This is the milestone that we have made about half(I wish...) of our works.
We make the `KvReader` use epoch-base concurrence, and succeed(maybe) to make a `build_index` with it.
TODO:
- Make writer write to right epoch.
- Rewrite `compact_file`, and figure out some method to collect garbage.
NOTE: This version cannot pass the tests. It probably shouldn't appear in the master branch,
however, considering this project isn't for production, that becomes acceptable.

Signed-off-by: YuJuncen <maruruku@stu.csust.edu.cn>